moved SeasonEpisodeCount to its own file
